# apps/auth/swagger.json
{"openapi":"3.0.0","paths":{"/api/auth/signup":{"post":{"operationId":"GatewayController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDto"}}}}},"tags":["ApiClient"]}},"/api/auth/signin":{"post":{"operationId":"GatewayController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDto"}}}}},"tags":["ApiClient"]}},"/api/auth/refreshToken":{"post":{"operationId":"GatewayController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDto"}}}}},"tags":["ApiClient"]}},"/api/auth/me":{"get":{"operationId":"GatewayController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["ApiClient"]}},"/api/generator/projects":{"post":{"operationId":"GatewayController_generatorCreateProject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreateDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectDto"}}}}},"tags":["ApiClient"]},"get":{"operationId":"GatewayController_generatorReadProjects","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectDto"}}}}}},"tags":["ApiClient"]}},"/api/generator/sync/{projectId}":{"post":{"operationId":"GatewayController_generatorSyncProject","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["ApiClient"]}},"/api/generator/website/{projectId}":{"get":{"operationId":"GatewayController_generatorWebsite","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ApiClient"]}}},"info":{"title":"NotionGen","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SignupDto":{"type":"object","properties":{"username":{"type":"string","minLength":4,"maxLength":16},"password":{"type":"string","pattern":"/((?=.*\\d)|(?=.*\\W+))(?![.\\n])(?=.*[A-Z])(?=.*[a-z]).*$/"}},"required":["username","password"]},"AuthUserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["id","username","accessToken","refreshToken"]},"SigninDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"RefreshTokenDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"AccessTokenDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["id","username","password"]},"ProjectCreateDto":{"type":"object","properties":{"notionId":{"type":"string"},"notionName":{"type":"string"},"notionAccessCode":{"type":"string"}},"required":["notionId","notionName","notionAccessCode"]},"ProjectDto":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"notionId":{"type":"string"},"notionName":{"type":"string"},"notionAccessCode":{"type":"string"},"createdAt":{"type":"number"}},"required":["userId","notionId","notionName","notionAccessCode"]}}}}